CameraFactory.java:247 Line Type mismatch:convert int to float err resolve;

+++ OpenDS3.0/src/eu/opends/camera/CameraFactory.java
@@ -244,7 +244,7 @@ public abstract class CameraFactory
     		mirrorModeString = "off";
     	this.mirrorMode = MirrorMode.valueOf(mirrorModeString.toUpperCase());
     	    	
-    	angleBetweenAdjacentCameras = settingsLoader.getSetting(Setting.General_angleBetweenAdjacentCameras, 40);
+    	angleBetweenAdjacentCameras = (float)settingsLoader.getSetting(Setting.General_angleBetweenAdjacentCameras, 40);
     	if(angleBetweenAdjacentCameras > 90 || angleBetweenAdjacentCameras < 0)
     	{
     		System.err.println("Angle between adjacent cameras must be within 0 to 90 degrees. Set to default: 40 degrees.");
